Abstractify PP counter logic from the "Triangles" implementation
